Lots of changes... - Core re-designed; - Lots of corrections on Presentation and DAOs - Minor bug fixes.
git-svn-id: svn+ufpa://linc.ufpa.br/var/lib/svn/trunk/XFlow@484 6cedf61c-d5b9-4d6e-8434-ec74007c3759

+++ src/br/ufpa/linc/xflow/metrics/entry/EntryMetricValues.java
@@ -102,7 +102,7 @@ public class EntryMetricValues extends MetricValuesTable {
 		this.author = author;
 	}
 
-	public int getValueByName(final String metricName) {
+	public double getValueByName(final String metricName) {
 		if(metricName == "Added Files"){
 			return this.getEntryAddedFiles();
 		}
